migrate class to separate file in a package
